Autoconverted sources with fixes ydanila/sharpen_imazen@1efc26385ba8512e1fc4b83cbca88ad92ccd4240 and updated configuration ydanila/j-metadata-extractor@d76165b62a18ac907115b24b30ca1c4d20b06517

+++ Com.Drew/Com/Drew/Metadata/Adobe/AdobeJpegDescriptor.cs
@@ -55,7 +55,7 @@ namespace Com.Drew.Metadata.Adobe
 		private string GetDctEncodeVersionDescription()
 		{
 			int? value = _directory.GetInteger(AdobeJpegDirectory.TagColorTransform);
-			return value == null ? null : value == unchecked((int)(0x64)) ? "100" : Sharpen.Extensions.ToString(value);
+			return value == null ? null : value == unchecked((int)(0x64)) ? "100" : Sharpen.Extensions.ToString((int)value);
 		}
 
 		private string GetColorTransformDescription()
