+++ Com.Drew/Com/Drew/Metadata/Directory.cs
@@ -884,7 +884,7 @@ namespace Com.Drew.Metadata
 			}
 			if (o is long?)
 			{
-				return new Rational((long?)o, 1);
+				return new Rational((long)(long?)o, 1);
 			}
 			// NOTE not doing conversions for real number types
 			return null;
